Signing off. Bugs remain. Update method is a murder scene. Known bugs:
Player doesn't let go of wall when grab is released. Jumping in neutral releases (because of jump). Moving away from wall releases. Moving into wall does nothing (as expected).

Jump does not work when grab is held. Moving away from wall works when grab is held, possibly because this turns the player and is therefore not considered to be touching a wall.